requirements.txt
requests
dnspython
whois